requirements.txt
# Core
flask>=3.0.0
pdfplumber>=0.10.0        # Native PDF text extraction (tables + text)
pymupdf>=1.23.0           # PDF fallback + scanned PDF rendering (import as fitz)
Pillow>=10.0.0            # Image loading and preprocessing
pytesseract>=0.3.10       # Python wrapper for Tesseract OCR engine

# ML (optional – needed for MLClassifier / EnsembleClassifier)
scikit-learn>=1.4.0       # TF-IDF + Logistic Regression
joblib>=1.3.0             # Model serialisation

# Transformer-based (optional upgrade path)
# transformers>=4.38.0
# torch>=2.0.0

# Testing
pytest>=7.4.0

# ─── System dependency (install separately) ────────────────────────────────
# Tesseract OCR engine:
#   Ubuntu/Debian : sudo apt install tesseract-ocr tesseract-ocr-eng
#   macOS         : brew install tesseract
#   Windows       : https://github.com/UB-Mannheim/tesseract/wiki
